+++ src/ladb_opencutlist/js/plugins/jquery.ladb.tab-importer.js
@@ -328,7 +328,7 @@
                             that.model_is_empty = false;
 
                             // Success notification
-                            that.opencutlist.notify(i18next.t('tab.importer.success.imported', {count: response.imported_part_count}), 'success', [
+                            that.opencutlist.notify(i18next.t('tab.importer.success.imported', { count: response.imported_part_count }), 'success', [
                                 Noty.button(i18next.t('default.see'), 'btn btn-default', function () {
                                     that.opencutlist.minimize();
                                 }),